frontend/requirements.txt
streamlit>=1.35.0
pandas
httpx
loguru
python-dotenv
markdown
pdfkit
streamlit-echarts
